Reworked the personal account layout and renamed some filed
Because:
- Java classes should start with a capital letter
- Classes that are related to each other should be grouped in the same package

In preparation for the mvvm restructure of the personal account
