+++ app/src/main/res/layout/fragment_own_study.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".views.ownStudyFragment"
+    tools:context=".views.ownStudies.OwnStudyFragment"
 
     app:layout_behavior="@string/appbar_scrolling_view_behavior">
